Correct naming and functions of tests
